+++ app/models/user.rb
@@ -24,7 +24,6 @@ class User < ActiveRecord::Base
   has_many :agreements, -> { where deleted: false }
   has_many :datasets, -> { where deleted: false }
   has_many :dataset_file_audits
-  has_many :lists
   has_many :tools
 
   # User Methods